requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.24,<1.0
pydantic>=2.4,<3.0
pyyaml>=6.0
python-kasa>=0.6.3
tinytuya>=1.12.0
lifxlan>=1.2.8,<1.3
zeroconf>=0.132.2
httpx>=0.24,<0.28
